+++ backend/internal/infra/ent/donorresponse.go
@@ -26,7 +26,7 @@ type DonorResponse struct {
 	// DeletedAt holds the value of the "deleted_at" field.
 	DeletedAt *time.Time `json:"deletedAt"`
 	// Amount holds the value of the "amount" field.
-	Amount int32 `json:"amount,omitempty"`
+	Amount float64 `json:"amount,omitempty"`
 	// CompensationType holds the value of the "compensation_type" field.
 	CompensationType donorresponse.CompensationType `json:"compensation_type,omitempty"`
 	// TaxiCompensation holds the value of the "taxi_compensation" field.
@@ -86,7 +86,7 @@ func (*DonorResponse) scanValues(columns []string) ([]any, error) {
 		case donorresponse.FieldTaxiCompensation, donorresponse.FieldIsConfirmed:
 			values[i] = new(sql.NullBool)
 		case donorresponse.FieldAmount:
-			values[i] = new(sql.NullInt64)
+			values[i] = new(sql.NullFloat64)
 		case donorresponse.FieldID, donorresponse.FieldCompensationType, donorresponse.FieldStatus:
 			values[i] = new(sql.NullString)
 		case donorresponse.FieldCreatedAt, donorresponse.FieldUpdatedAt, donorresponse.FieldDeletedAt:
@@ -136,10 +136,10 @@ func (_m *DonorResponse) assignValues(columns []string, values []any) error {
 				*_m.DeletedAt = value.Time
 			}
 		case donorresponse.FieldAmount:
-			if value, ok := values[i].(*sql.NullInt64); !ok {
+			if value, ok := values[i].(*sql.NullFloat64); !ok {
 				return fmt.Errorf("unexpected type %T for field amount", values[i])
 			} else if value.Valid {
-				_m.Amount = int32(value.Int64)
+				_m.Amount = value.Float64
 			}
 		case donorresponse.FieldCompensationType:
 			if value, ok := values[i].(*sql.NullString); !ok {
